Remove Python 2 and Jython support
