+++ atest/get_connection.robot
@@ -1,5 +1,4 @@
 *** Settings ***
-Default Tags   pybot   jybot
 Resource       resources/common.robot
 Test Teardown  Close All Connections
 Library        OperatingSystem
