+++ atest/put_directory.robot
@@ -1,5 +1,4 @@
 *** Settings ***
-Default Tags    pybot   jybot
 Resource        resources/sftp.robot
 Suite Setup     Login As Valid User
 Suite Teardown  Close All Connections
@@ -31,7 +30,6 @@ Put Directory Including Subdirectories To Existing Remote Path With SCP (transfe
     [Teardown]  Execute Command  rm -rf ./textfiles
 
 Put Directory Including Subdirectories To Existing Remote Path With SCP (all)
-    [Tags]  pybot
     Put Directory  ${LOCAL TEXTFILES}  .  recursive=True  scp=ALL
     Remote Directory Should Exist With Subdirectories  ./textfiles
     [Teardown]  Execute Command  rm -rf ./textfiles
@@ -83,7 +81,6 @@ Put Directory And Check For Proper Permissions
 	[Teardown]  Execute Command  rm -rf ${CURDIR}${/}testdata${/}to_put
 
 Put Directory With SCP (transfer) And Preserve Time
-    [Tags]  pybot
     ${current_time} =  Get Current Date  result_format=epoch  exclude_millis=True
     Put Directory  ${LOCAL TEXTFILES}  .  recursive=True  scp=TRANSFER  scp_preserve_times=True
     Remote Directory Should Exist With Subdirectories  ./textfiles
@@ -94,7 +91,6 @@ Put Directory With SCP (transfer) And Preserve Time
     [Teardown]  Execute Command  rm -rf ./textfiles
 
 Put Directory With SCP (all) And Preserve Time
-    [Tags]  pybot
     ${current_time} =  Get Current Date  result_format=epoch  exclude_millis=True
     Put Directory  ${LOCAL TEXTFILES}  .  recursive=True  scp=ALL  scp_preserve_times=True
     ${last_access_time} =  Execute Command  stat -c %X ./textfiles
