+++ atest/write_and_read/other_writes_and_reads.robot
@@ -1,6 +1,5 @@
 *** Settings ***
 Resource        ../resources/write_and_read_resource.robot
-Default Tags      pybot   jybot
 Suite Setup     Run Keywords  Login And Upload Test Scripts  AND  Put File  ${CORRUPTED FILE}  ${REMOTE TEST ROOT}
 Suite Teardown  Remove Test Files and Close Connections
 
@@ -93,7 +92,6 @@ Read Until REGEXP Prompt With Strip Prompt
     [Teardown]                      Set Client Configuration         prompt=${PROMPT}
 
 Configure Session Width And Height
-    [Tags]   pybot
     Set Client Configuration  prompt=${PROMPT}  height=48  width=160
     ${conn}  Get Connection  1
     Should Be Equal As Integers  ${conn.height}  48
@@ -101,12 +99,6 @@ Configure Session Width And Height
     Write  stty size
     ${output}=  Read Until Prompt
     Should Contain  ${output}  48 160
-    [Teardown]  Set Client Configuration  height=24  width=80
-
-Configure Session Width And Height Not Supported
-    [Tags]   jybot
-    [Documentation]  WARN  1.1:1  Setting width or height is not supported with Jython.
-    Set Client Configuration  prompt=${PROMPT}  height=48  width=160
     [Teardown]  Set Client Configuration  height=24  width=80
 
 Read Until With Encoding Errors On Strict
