+++ requirements-dev.txt
@@ -1,3 +1,3 @@
-robotframework >= 4.0
+robotframework >= 3.0
 robotstatuschecker == 1.5.1
 robotbackgroundlogger >= 1.2
