+++ src/SSHLibrary/library.py
@@ -16,23 +16,23 @@
 from __future__ import print_function
 
 import re
-import os
 from .deco import keyword
+
 try:
     from robot.api import logger
 except ImportError:
     logger = None
 
+from robot.utils import is_string, is_truthy, plural_or_not
 from .sshconnectioncache import SSHConnectionCache
-from .abstractclient import SSHClientException
+from .client import SSHClientException
 from .client import SSHClient
 from .config import (Configuration, IntegerEntry, LogLevelEntry, NewlineEntry,
                      StringEntry, TimeEntry)
-from .utils import ConnectionCache, is_string, is_truthy, plural_or_not
 from .version import VERSION
 
-
 __version__ = VERSION
+
 
 class SSHLibrary(object):
     """SSHLibrary is a Robot Framework test library for SSH and SFTP.
@@ -612,7 +612,7 @@ class SSHLibrary(object):
         *Note:* This keyword does not work when using Jython.
         """
         if SSHClient.enable_logging(logfile):
-            self._log('SSH log is written to <a href="%s">file</a>.' % logfile,
+            self._log(f'SSH log is written to <a href="{logfile}">file</a>.',
                       'HTML')
 
     def open_connection(self, host, alias=None, port=22, timeout=None,
@@ -745,7 +745,6 @@ class SSHLibrary(object):
         connections = self._connections
         connections.close_current()
 
-
     def close_all_connections(self):
         """Closes all open connections.
 
@@ -876,7 +875,7 @@ class SSHLibrary(object):
             if logger:
                 logger.write(msg, level)
             else:
-                print('*%s* %s' % (level, msg))
+                print(f'*{level}* {msg}')
 
     def _active_loglevel(self, level):
         if level is None:
@@ -884,7 +883,7 @@ class SSHLibrary(object):
         if is_string(level) and \
                 level.upper() in ['TRACE', 'DEBUG', 'INFO', 'WARN', 'HTML', 'NONE']:
             return level.upper()
-        raise AssertionError("Invalid log level '%s'." % level)
+        raise AssertionError(f"Invalid log level '{level}'.")
 
     def _get_config_values(self, config, index, host, alias, port, timeout,
                            newline, prompt, term_type, width, height, encoding, escape_ansi):
@@ -1001,7 +1000,7 @@ class SSHLibrary(object):
         """
         jumphost_connection_conf = self.get_connection(index_or_alias=jumphost_index_or_alias) \
             if jumphost_index_or_alias else None
-        jumphost_connection = self._connections.connections[jumphost_connection_conf.index-1] \
+        jumphost_connection = self._connections.connections[jumphost_connection_conf.index - 1] \
             if jumphost_connection_conf and jumphost_connection_conf.index else None
 
         return self._login(self.current.login, username, password, is_truthy(allow_agent),
@@ -1072,22 +1071,24 @@ class SSHLibrary(object):
         """
         if proxy_cmd and jumphost_index_or_alias:
             raise ValueError("`proxy_cmd` and `jumphost_connection` are mutually exclusive SSH features.")
-        jumphost_connection_conf = self.get_connection(index_or_alias=jumphost_index_or_alias) if jumphost_index_or_alias else None
-        jumphost_connection = self._connections.connections[jumphost_connection_conf.index-1] if jumphost_connection_conf and jumphost_connection_conf.index else None
+        jumphost_connection_conf = self.get_connection(
+            index_or_alias=jumphost_index_or_alias) if jumphost_index_or_alias else None
+        jumphost_connection = self._connections.connections[
+            jumphost_connection_conf.index - 1] if jumphost_connection_conf and jumphost_connection_conf.index else None
         return self._login(self.current.login_with_public_key, username,
                            keyfile, password, is_truthy(allow_agent),
                            is_truthy(look_for_keys), delay, proxy_cmd,
                            jumphost_connection, is_truthy(read_config), keep_alive_interval)
 
     def _login(self, login_method, username, *args):
-        self._log("Logging into '%s:%s' as '%s'."
-                   % (self.current.config.host, self.current.config.port,
-                      username), self._config.loglevel)
+        self._log(
+            f"Logging into '{self.current.config.host}:{self.current.config.port}' as '{self.current.config.host}'.",
+            self._config.loglevel)
         try:
             login_output = login_method(username, *args)
             if is_truthy(self.current.config.escape_ansi):
                 login_output = self._escape_ansi_sequences(login_output)
-            self._log('Read output: %s' % login_output, self._config.loglevel)
+            self._log(f'Read output: {login_output}', self._config.loglevel)
             return login_output
         except SSHClientException as e:
             raise RuntimeError(e)
@@ -1125,7 +1126,7 @@ class SSHLibrary(object):
         return banner.decode(self.DEFAULT_ENCODING)
 
     def execute_command(self, command, return_stdout=True, return_stderr=False,
-                        return_rc=False, sudo=False,  sudo_password=None, timeout=None, output_during_execution=False,
+                        return_rc=False, sudo=False, sudo_password=None, timeout=None, output_during_execution=False,
                         output_if_timeout=False, invoke_subsystem=False, forward_agent=False):
         """Executes ``command`` on the remote machine and returns its outputs.
 
@@ -1182,7 +1183,7 @@ class SSHLibrary(object):
         ``command`` argument. If the server allows it, the channel will then be
         directly connected to the requested subsystem.
 
-        ``forward_agent`` determines whether to forward the local SSH Agent process to the process being executed. 
+        ``forward_agent`` determines whether to forward the local SSH Agent process to the process being executed.
         This assumes that there is an agent in use (i.e. `eval $(ssh-agent)`). Setting ``forward_agent`` does not
         work with Jython.
         | `Execute Command` | ssh-add -L | forward_agent=True |
@@ -1197,9 +1198,9 @@ class SSHLibrary(object):
         ``output_during_execution`` and ``output_if_timeout`` are not working with Jython. New in SSHLibrary 3.5.0.
         """
         if not is_truthy(sudo):
-            self._log("Executing command '%s'." % command, self._config.loglevel)
+            self._log(f"Executing command '{command}'.", self._config.loglevel)
         else:
-            self._log("Executing command 'sudo %s'." % command, self._config.loglevel)
+            self._log(f"Executing command 'sudo {command}'.", self._config.loglevel)
         opts = self._legacy_output_options(return_stdout, return_stderr,
                                            return_rc)
         stdout, stderr, rc = self.current.execute_command(command, sudo, sudo_password,
@@ -1207,7 +1208,7 @@ class SSHLibrary(object):
                                                           is_truthy(invoke_subsystem), forward_agent)
         return self._return_command_output(stdout, stderr, rc, *opts)
 
-    def start_command(self, command, sudo=False,  sudo_password=None, invoke_subsystem=False, forward_agent=False):
+    def start_command(self, command, sudo=False, sudo_password=None, invoke_subsystem=False, forward_agent=False):
         """Starts execution of the ``command`` on the remote machine and returns immediately.
 
         This keyword returns nothing and does not wait for the ``command``
@@ -1251,9 +1252,9 @@ class SSHLibrary(object):
         ``invoke_subsystem`` is new in SSHLibrary 3.4.0.
         """
         if not is_truthy(sudo):
-            self._log("Starting command '%s'." % command, self._config.loglevel)
+            self._log(f"Starting command '{command}'.", self._config.loglevel)
         else:
-            self._log("Starting command 'sudo %s'." % command, self._config.loglevel)
+            self._log(f"Starting command 'sudo {command}'.", self._config.loglevel)
         if self.current.config.index not in self._last_commands.keys():
             self._last_commands[self.current.config.index] = command
         else:
@@ -1310,7 +1311,8 @@ class SSHLibrary(object):
 
         This keyword logs the read command with log level ``INFO``.
         """
-        self._log("Reading output of command '%s'." % self._last_commands.get(self.current.config.index), self._config.loglevel)
+        self._log(f"Reading output of command '{self._last_commands.get(self.current.config.index)}'.",
+                  self._config.loglevel)
         opts = self._legacy_output_options(return_stdout, return_stderr,
                                            return_rc)
         try:
@@ -1363,7 +1365,7 @@ class SSHLibrary(object):
 
     def _return_command_output(self, stdout, stderr, rc, return_stdout,
                                return_stderr, return_rc):
-        self._log("Command exited with return code %d." % rc, self._config.loglevel)
+        self._log(f"Command exited with return code {rc}.", self._config.loglevel)
         ret = []
         if is_truthy(return_stdout):
             ret.append(stdout.rstrip('\n'))
@@ -1595,7 +1597,7 @@ class SSHLibrary(object):
     def _escape_ansi_sequences(output):
         ansi_escape = re.compile(r'(?:\x1B[@-_]|[\x80-\x9F])[0-?]*[ -/]*[@-~]', flags=re.IGNORECASE)
         output = ansi_escape.sub('', output)
-        return ("%r" % output)[1:-1].encode().decode('unicode-escape')
+        return (f"{output!r}")[1:-1].encode().decode('unicode-escape')
 
     def get_file(self, source, destination='.', scp='OFF', scp_preserve_times=False):
         """Downloads file(s) from the remote machine to the local machine.
@@ -1828,7 +1830,7 @@ class SSHLibrary(object):
             raise RuntimeError(e)
         if files:
             for src, dst in files:
-                self._log("'%s' -> '%s'" % (src, dst), self._config.loglevel)
+                self._log(f"'{src}' -> '{dst}'", self._config.loglevel)
 
     def file_should_exist(self, path):
         """Fails if the given ``path`` does NOT point to an existing file.
@@ -1843,7 +1845,7 @@ class SSHLibrary(object):
         | `File Should Exist` | /initrd.img | # Points to /boot/initrd.img |
         """
         if not self.current.is_file(path):
-            raise AssertionError("File '%s' does not exist." % path)
+            raise AssertionError(f"File '{path}' does not exist.")
 
     def file_should_not_exist(self, path):
         """Fails if the given ``path`` points to an existing file.
@@ -1858,7 +1860,7 @@ class SSHLibrary(object):
         ``/non/existing`` is a link that points an existing file.
         """
         if self.current.is_file(path):
-            raise AssertionError("File '%s' exists." % path)
+            raise AssertionError(f"File '{path}' exists.")
 
     def directory_should_exist(self, path):
         """Fails if the given ``path`` does not point to an existing directory.
@@ -1873,7 +1875,7 @@ class SSHLibrary(object):
         | `Directory Should Exist` | /usr/local/man | # Points to /usr/share/man/ |
         """
         if not self.current.is_dir(path):
-            raise AssertionError("Directory '%s' does not exist." % path)
+            raise AssertionError(f"Directory '{path}' does not exist.")
 
     def directory_should_not_exist(self, path):
         """Fails if the given ``path`` points to an existing directory.
@@ -1888,7 +1890,7 @@ class SSHLibrary(object):
         ``/non/existing`` is a link that points to an existing directory.
         """
         if self.current.is_dir(path):
-            raise AssertionError("Directory '%s' exists." % path)
+            raise AssertionError(f"Directory '{path}' exists.")
 
     def list_directory(self, path, pattern=None, absolute=False):
         """Returns and logs items in the remote ``path``, optionally filtered with ``pattern``.
@@ -1923,8 +1925,8 @@ class SSHLibrary(object):
             items = self.current.list_dir(path, pattern, is_truthy(absolute))
         except SSHClientException as msg:
             raise RuntimeError(msg)
-        self._log('%d item%s:\n%s' % (len(items), plural_or_not(items),
-                                       '\n'.join(items)), self._config.loglevel)
+        self._log("{0} item{1}:\n{2}".format(len(items), plural_or_not(items),
+                                             '\n'.join(items)), self._config.loglevel)
         return items
 
     def list_files_in_directory(self, path, pattern=None, absolute=False):
@@ -1935,8 +1937,8 @@ class SSHLibrary(object):
         except SSHClientException as msg:
             raise RuntimeError(msg)
         files = self.current.list_files_in_dir(path, pattern, absolute)
-        self._log('%d file%s:\n%s' % (len(files), plural_or_not(files),
-                                       '\n'.join(files)), self._config.loglevel)
+        self._log('{0} file{1}:\n{2}'.format(len(files), plural_or_not(files),
+                                             '\n'.join(files)), self._config.loglevel)
         return files
 
     def list_directories_in_directory(self, path, pattern=None, absolute=False):
@@ -1945,9 +1947,9 @@ class SSHLibrary(object):
             dirs = self.current.list_dirs_in_dir(path, pattern, is_truthy(absolute))
         except SSHClientException as msg:
             raise RuntimeError(msg)
-        self._log('%d director%s:\n%s' % (len(dirs),
-                                           'y' if len(dirs) == 1 else 'ies',
-                                           '\n'.join(dirs)), self._config.loglevel)
+        self._log('{0} director{1}:\n{2}'.format(len(dirs),
+                                                 'y' if len(dirs) == 1 else 'ies',
+                                                 '\n'.join(dirs)), self._config.loglevel)
         return dirs
