--- src/SSHLibrary/utils.py
@@ -1,24 +1,0 @@
-#  Copyright 2008-2015 Nokia Networks
-#  Copyright 2016-     Robot Framework Foundation
-#
-#  Licensed under the Apache License, Version 2.0 (the "License");
-#  you may not use this file except in compliance with the License.
-#  You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-#  Unless required by applicable law or agreed to in writing, software
-#  distributed under the License is distributed on an "AS IS" BASIS,
-#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-#  See the License for the specific language governing permissions and
-#  limitations under the License.
-
-from robot.utils import (is_bytes, is_list_like, is_string, is_truthy,
-                         is_unicode, plural_or_not, secs_to_timestr,
-                         timestr_to_secs, ConnectionCache)
-
-
-try:
-    unicode = unicode
-except NameError:
-    unicode = str
